Factor font stuff into its own module
