+++ Dark Samurai Warrior Reloaded/Dark Samurai Warrior Reloaded.vcxproj
@@ -129,6 +129,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="font\font.c" />
     <ClCompile Include="gui\gui.c" />
     <ClCompile Include="input\input.c" />
     <ClCompile Include="main.c" />
@@ -140,6 +141,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="common.h" />
+    <ClInclude Include="font\font.h" />
     <ClInclude Include="gui\gui.h" />
     <ClInclude Include="input\input.h" />
     <ClInclude Include="math.h" />
