+++ Dark Samurai Warrior Reloaded/gui/gui.h
@@ -3,7 +3,7 @@
 
 #include "../common.h"
 #include "../math.h"
-#include "../render.h"
+#include "../font/font.h"
 
 typedef V4 Color;
 typedef u64 ID;
